Changes to storage writer reader interface #3024
