+++ tests/analysis/hash_tagging.py
@@ -7,6 +7,7 @@ import unittest
 from dfvfs.path import fake_path_spec
 
 from plaso.analysis import hash_tagging
+from plaso.containers import events
 from plaso.lib import definitions
 
 from tests.analysis import test_lib
@@ -108,7 +109,8 @@ class HashTaggingAnalysisPluginTest(test_lib.AnalysisPluginTestCase):
     self.assertEqual(report.text, expected_text)
 
     labels = []
-    for event_tag in storage_writer.GetEventTags():
+    for event_tag in storage_writer.GetAttributeContainers(
+        events.EventTag.CONTAINER_TYPE):
       labels.extend(event_tag.labels)
     self.assertEqual(len(labels), 1)
 
@@ -132,7 +134,8 @@ class HashTaggingAnalysisPluginTest(test_lib.AnalysisPluginTestCase):
     self.assertEqual(report.text, expected_text)
 
     labels = []
-    for event_tag in storage_writer.GetEventTags():
+    for event_tag in storage_writer.GetAttributeContainers(
+        events.EventTag.CONTAINER_TYPE):
       labels.extend(event_tag.labels)
     self.assertEqual(len(labels), 0)
